netboot: mount / read-only
This should solve all the problem where services get started on the contestant
machine but failing because the '/' is read-only.

The systemd services are using 'ConditionNeedsUpdate=/etc'. This predicate is
checking if '/etc' is read-write but without the 'ro' option, the system was
booted with '/' mounted as 'rw'. Thus, systemd believed '/etc' was writeable
and when in fact the writes where denied by the nfs-server because exported as
'ro'.

With this patch, '/' is mounted 'ro' and the services are not started.