xrun: simulate a PTY input for machinectl
Machinectl stops if its standard input is closed because it always
expects to be run from a PTY, and there seems to be no --no-pty option.

Systemd upstream suggests to use systemd-run instead
(https://github.com/systemd/systemd/issues/5042), but it does not work
for our use case as we need the session opened by machinectl to be able
to run a service from the controller of the user we connect to.

So we just feed it a `tail -f /dev/null` which will never close its input
but will not flood its output with garbage like `yes` does.