masternode: refactor to be stateless (#180)
masternode now queries the DB periodically to get the list of tasks and does not maintain that state in memory.

This enable DB updates that are reflected in masternode.

Known issue: this will result in heavy load on the DB if there is a large amount of pending tasks returned at each query. This will be mitigated in the future change with rate limiting.

* concours: add "failed" task status

This status means that masternode could not run the task successfully.