ansible: replace sddm with lightdm & lightdm-prologin-greeter.
This removes presenced/client.py which was doing very convoluted HTTP
requests to call a local function (!).

This removes useless config.

This replaces presenced.server with a much simpler 'heartbeat sender'
that only sends periodic signals to PresenceSync.

This refactors pam_presenced into a multi-hook pam_prologin, which does
basically the same things, but does them in the 'account' stage, so we
can get interactive messages sent to the user during the
potentially-blocking operations such as HFS moves.